move validation from ArnValidation object to QueueArnValidation
